requirements.txt
scanpy>=1.9
anndata>=0.10
torch
numpy
scipy
pandas
PyYAML
wandb
